factored out request function
